update make file and python files to run from the make
