+++ models/ngspice/testing/regression/bjt_beta/models_regression.py
@@ -1,3 +1,16 @@
+# Copyright 2022 GlobalFoundries PDK Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """
 Usage:
   models_regression.py [--num_cores=<num>]
@@ -7,8 +20,6 @@ Usage:
   --num_cores=<num>      Number of cores to be used by simulator
 """
 
-from cmath import inf
-from re import T
 from docopt import docopt
 import pandas as pd
 import numpy as np
@@ -16,9 +27,22 @@ import os
 from jinja2 import Template
 import concurrent.futures
 import shutil
+import multiprocessing as mp
+
+import glob
+
 import warnings
 
 warnings.simplefilter(action="ignore", category=FutureWarning)
+
+PASS_THRESH = 2.0
+
+
+def find_bjt(filepath):
+    """
+    Find bjt in csv files
+    """
+    return os.path.exists(filepath)
 
 
 def call_simulator(file_name):
@@ -26,312 +50,342 @@ def call_simulator(file_name):
     Args:
         file_name (str): Netlist file name.
     """
-    os.system(f"ngspice -b -a {file_name} -o {file_name}.log > {file_name}.log")
+    return os.system(
+        f"ngspice -b -a {file_name} -o {file_name}.log > {file_name}.log"
+        )
 
 
-def ext_measured(device, vb, step, Id_sim, list_devices, vc):
+def ext_npn_measured(dev_data_path, device, devices, dev_path):
+    # Read Data
+    df = pd.read_excel(dev_data_path)
 
-    # Get dimensions used for each device
-    dimensions = pd.read_csv(f"{device}/{device}.csv", usecols=["corners"])
-    loops = dimensions["corners"].count()
+    all_dfs = []
 
-    # Extracting measured values for each Device
+    loops = df["corners"].count()
+
     for i in range(loops):
+
+        temp_range = int(loops / 4)
+        if i in range(0, temp_range):
+            temp = 25
+        elif i in range(temp_range, 2 * temp_range):
+            temp = -40
+        elif i in range(2 * temp_range, 3 * temp_range):
+            temp = 125
+        else:
+            temp = 175
+
+        tempr = []
+        dev = []
+        ic_meas = []
+        ib_meas = []
+
         k = i
-        if i >= len(list_devices):
-            while k >= len(list_devices):
-                k = k - len(list_devices)
+        if i >= len(devices):
+            while k >= len(devices):
+                k = k - len(devices)
 
         # Special case for 1st measured values
+
         if i == 0:
-            if device == "pnp":
-                temp_vb = vb
-                vb = "-vb "
-            # measured Id_sim 0
-            col_list = [f"{vb}", f"{vc}{step[0]}", f"{vc}{step[1]}", f"{vc}{step[2]}"]
-            df_measured = pd.read_csv(f"{device}/{device}.csv", usecols=col_list)
-            df_measured.columns = [
-                f"{vb}",
-                f"{vc}{step[0]}",
-                f"{vc}{step[1]}",
-                f"{vc}{step[2]}",
-            ]
-            df_measured.to_csv(
-                f"{device}/measured_{Id_sim[0]}/{i}_measured_{list_devices[k]}.csv",
-                index=False,
+
+            idf_ic = df[["vbp ", "vcp =1", "vcp =2", "vcp =3"]].copy()
+            idf_ic.rename(
+                columns={
+                    "vbp ": "measured_base_volt",
+                    "vcp =1": "measured_ic_vcp_step1",
+                    "vcp =2": "measured_ic_vcp_step2",
+                    "vcp =3": "measured_ic_vcp_step3",
+                },
+                inplace=True,
             )
 
-            if device == "pnp":
-                vb = temp_vb
-
-            # measured Id_sim 1
-            col_list = [
-                f"{vb}",
-                f"{vc}{step[0]}.{2*i+1}",
-                f"{vc}{step[1]}.{2*i+1}",
-                f"{vc}{step[2]}.{2*i+1}",
-            ]
-            df_measured = pd.read_csv(f"{device}/{device}.csv", usecols=col_list)
-            df_measured.columns = [
-                f"{vb}",
-                f"{vc}{step[0]}",
-                f"{vc}{step[1]}",
-                f"{vc}{step[2]}",
-            ]
-            df_measured.to_csv(
-                f"{device}/measured_{Id_sim[1]}/{i}_measured_{list_devices[k]}.csv",
-                index=False,
-            )
         else:
-            # measured Id_sim 0
-            col_list = [
-                f"{vb}",
-                f"{vc}{step[0]}.{2*i}",
-                f"{vc}{step[1]}.{2*i}",
-                f"{vc}{step[2]}.{2*i}",
-            ]
-            df_measured = pd.read_csv(f"{device}/{device}.csv", usecols=col_list)
-            df_measured.columns = [
-                f"{vb}",
-                f"{vc}{step[0]}",
-                f"{vc}{step[1]}",
-                f"{vc}{step[2]}",
-            ]
-            df_measured.to_csv(
-                f"{device}/measured_{Id_sim[0]}/{i}_measured_{list_devices[k]}.csv",
-                index=False,
+
+            idf_ic = df[
+                ["vbp ", f"vcp =1.{2*i}", f"vcp =2.{2*i}", f"vcp =3.{2*i}"]
+            ].copy()
+            idf_ic.rename(
+                columns={
+                    "vbp ": "measured_base_volt",
+                    f"vcp =1.{2*i}": "measured_ic_vcp_step1",
+                    f"vcp =2.{2*i}": "measured_ic_vcp_step2",
+                    f"vcp =3.{2*i}": "measured_ic_vcp_step3",
+                },
+                inplace=True,
             )
 
-            # measured Id_sim 1
-            col_list = [
-                f"{vb}",
-                f"{vc}{step[0]}.{2*i+1}",
-                f"{vc}{step[1]}.{2*i+1}",
-                f"{vc}{step[2]}.{2*i+1}",
-            ]
-            df_measured = pd.read_csv(f"{device}/{device}.csv", usecols=col_list)
-            df_measured.columns = [
-                f"{vb}",
-                f"{vc}{step[0]}",
-                f"{vc}{step[1]}",
-                f"{vc}{step[2]}",
-            ]
-            df_measured.to_csv(
-                f"{device}/measured_{Id_sim[1]}/{i}_measured_{list_devices[k]}.csv",
-                index=False,
+        os.makedirs(f"{dev_path}/ic_measured", exist_ok=True)
+        idf_ic.to_csv(
+            f"{dev_path}/ic_measured/measured_{devices[k]}_t{temp}.csv"
             )
 
-
-def ext_simulated(device, vc, step, sweep, Id_sim, list_devices, ib):
-
-    # Get dimensions used for each device
-    dimensions = pd.read_csv(f"{device}/{device}.csv", usecols=["corners"])
-    loops = dimensions["corners"].count()
-    temp_range = int(loops / 4)
-    netlist_tmp = f"./device_netlists/{device}.spice"
-    for i in range(loops):
-        if i in range(0, temp_range):
-            temp = 25
-        elif i in range(temp_range, 2 * temp_range):
-            temp = -40
-        elif i in range(2 * temp_range, 3 * temp_range):
-            temp = 125
-        else:
-            temp = 175
-
-        k = i
-        if i >= len(list_devices):
-            while k >= len(list_devices):
-                k = k - len(list_devices)
-
-        with open(netlist_tmp) as f:
-            tmpl = Template(f.read())
-            os.makedirs(f"{device}/{device}_netlists_sim", exist_ok=True)
-            with open(
-                f"{device}/{device}_netlists_sim/{i}_{device}_netlist_{list_devices[k]}.spice",
-                "w",
-            ) as netlist:
-                netlist.write(tmpl.render(device=list_devices[k], i=i, temp=temp))
-            netlist_path = f"{device}/{device}_netlists_sim/{i}_{device}_netlist_{list_devices[k]}.spice"
-
-            # Running ngspice for each netlist
-            with concurrent.futures.ProcessPoolExecutor(
-                max_workers=workers_count
-            ) as executor:
-                executor.submit(call_simulator, netlist_path)
-
-        # Writing simulated data 0
-        df_simulated = pd.read_csv(
-            f"{device}/simulated_{Id_sim[0]}/{i}_simulated_{list_devices[k]}.csv",
-            header=None,
-            delimiter=r"\s+",
+        idf_ib = df[
+            ["vbp ", f"vcp =1.{2*i+1}", f"vcp =2.{2*i+1}", f"vcp =3.{2*i+1}"]
+        ].copy()
+        idf_ib.rename(
+            columns={
+                "vbp ": "measured_base_volt",
+                f"vcp =1.{2*i+1}": "measured_ib_vcp_step1",
+                f"vcp =2.{2*i+1}": "measured_ib_vcp_step2",
+                f"vcp =3.{2*i+1}": "measured_ib_vcp_step3",
+            },
+            inplace=True,
         )
 
-        # empty array to append in it shaped (sweep, number of trials + 1)
-        new_array = np.empty((sweep, 1 + int(df_simulated.shape[0] / sweep)))
-        new_array[:, 0] = df_simulated.iloc[:sweep, 0]
-        times = int(df_simulated.shape[0] / sweep)
+        os.makedirs(f"{dev_path}/ib_measured", exist_ok=True)
+        idf_ib.to_csv(
+            f"{dev_path}/ib_measured/measured_{devices[k]}_t{temp}.csv"
+            )
 
-        for j in range(times):
-            new_array[:, (j + 1)] = df_simulated.iloc[j * sweep : (j + 1) * sweep, 1]
+        dev.append(devices[k])
+        tempr.append(temp)
+        ic_meas.append(
+            f"{dev_path}/ic_measured/measured_{devices[k]}_t{temp}.csv"
+            )
+        ib_meas.append(
+            f"{dev_path}/ib_measured/measured_{devices[k]}_t{temp}.csv"
+            )
 
-        # Writing final simulated data 0
-        df_simulated = pd.DataFrame(new_array)
-        df_simulated.to_csv(
-            f"{device}/simulated_{Id_sim[0]}/{i}_simulated_{list_devices[k]}.csv",
-            index=False,
-        )
-        df_simulated.columns = [
-            f"{vc}",
-            f"{ib}{step[0]}",
-            f"{ib}{step[1]}",
-            f"{ib}{step[2]}",
-        ]
-        df_simulated.to_csv(
-            f"{device}/simulated_{Id_sim[0]}/{i}_simulated_{list_devices[k]}.csv",
-            index=False,
-        )
-
-        # Writing simulated data 1
-        df_simulated = pd.read_csv(
-            f"{device}/simulated_{Id_sim[1]}/{i}_simulated_{list_devices[k]}.csv",
-            header=None,
-            delimiter=r"\s+",
-        )
-
-        # empty array to append in it shaped (sweep, number of trials + 1)
-        new_array = np.empty((sweep, 1 + int(df_simulated.shape[0] / sweep)))
-        new_array[:, 0] = df_simulated.iloc[:sweep, 0]
-        times = int(df_simulated.shape[0] / sweep)
-
-        for j in range(times):
-            new_array[:, (j + 1)] = df_simulated.iloc[j * sweep : (j + 1) * sweep, 1]
-
-        # Writing final simulated data 1
-        df_simulated = pd.DataFrame(new_array)
-        df_simulated.to_csv(
-            f"{device}/simulated_{Id_sim[1]}/{i}_simulated_{list_devices[k]}.csv",
-            index=False,
-        )
-        df_simulated.columns = [
-            f"{vc}",
-            f"{ib}{step[0]}",
-            f"{ib}{step[1]}",
-            f"{ib}{step[2]}",
-        ]
-        df_simulated.to_csv(
-            f"{device}/simulated_{Id_sim[1]}/{i}_simulated_{list_devices[k]}.csv",
-            index=False,
-        )
-
-
-def error_cal(device, vb, step, Id_sim, list_devices, vc):
-
-    df_final = pd.DataFrame()
-    # Get dimensions used for each device
-    dimensions = pd.read_csv(f"{device}/{device}.csv", usecols=["corners"])
-    loops = dimensions["corners"].count()
-    temp_range = int(loops / 4)
-    for i in range(loops):
-        if i in range(0, temp_range):
-            temp = 25
-        elif i in range(temp_range, 2 * temp_range):
-            temp = -40
-        elif i in range(2 * temp_range, 3 * temp_range):
-            temp = 125
-        else:
-            temp = 175
-
-        k = i
-        if i >= len(list_devices):
-            while k >= len(list_devices):
-                k = k - len(list_devices)
-
-        measured = pd.read_csv(
-            f"{device}/measured_{Id_sim}/{i}_measured_{list_devices[k]}.csv"
-        )
-        simulated = pd.read_csv(
-            f"{device}/simulated_{Id_sim}/{i}_simulated_{list_devices[k]}.csv"
-        )
-
-        error_1 = round(
-            100
-            * abs(
-                (abs(measured.iloc[0:, 1]) - abs(simulated.iloc[0:, 1]))
-                / abs(measured.iloc[:, 1])
-            ),
-            6,
-        )
-        error_2 = round(
-            100
-            * abs(
-                (abs(measured.iloc[0:, 2]) - abs(simulated.iloc[0:, 2]))
-                / abs(measured.iloc[:, 2])
-            ),
-            6,
-        )
-        error_3 = round(
-            100
-            * abs(
-                (abs(measured.iloc[0:, 3]) - abs(simulated.iloc[0:, 3]))
-                / abs(measured.iloc[:, 3])
-            ),
-            6,
-        )
-
-        df_error = pd.DataFrame(
-            data=[measured.iloc[:, 0], error_1, error_2, error_3]
-        ).transpose()
-        df_error.replace(
-            [np.inf, -np.inf], df_error.max().nlargest(2).iloc[1], inplace=True
-        )
-        df_error.to_csv(
-            f"{device}/error_{Id_sim}/{i}_{device}_error_{list_devices[k]}.csv",
-            index=False,
-        )
-
-        # Mean error
-        mean_error = (
-            df_error[f"{vc}{step[0]}"].mean()
-            + df_error[f"{vc}{step[1]}"].mean()
-            + df_error[f"{vc}{step[2]}"].mean()
-        ) / 6
-        # Max error
-        max_error = (
-            df_error[[f"{vc}{step[0]}", f"{vc}{step[1]}", f"{vc}{step[2]}"]].max().max()
-        )
-        # Max error location
-        max_index = max((df_error == max_error).idxmax())
-        max_location_vc = (df_error == max_error).idxmax(axis=1)[max_index]
-        if Id_sim == "Ic":
-            if i == 0:
-                if device == "pnp":
-                    temp_vb = vb
-                    vb = "-vb "
-            else:
-                if device == "pnp":
-                    vb = temp_vb
-        max_location_vb = df_error[f"{vb}"][max_index]
-
-        df_final_ = {
-            "Run no.": f"{i}",
-            "Temp": f"{temp}",
-            "Device name": f"{device}",
-            "device": f"{list_devices[k]}",
-            "Simulated_Val": f"{Id_sim}",
-            "Mean error%": f'{"{:.2f}".format(mean_error)}',
-            "Max error%": f'{"{:.2f}".format(max_error)} @ {max_location_vc} & Vc (V) = {max_location_vb}',
+        sdf = {
+            "device": dev,
+            "temp": tempr,
+            "ic_measured": ic_meas,
+            "ib_measured": ib_meas,
         }
-        df_final = df_final.append(df_final_, ignore_index=True)
+        sdf = pd.DataFrame(sdf)
+        all_dfs.append(sdf)
 
-    # Max mean error
-    print(df_final)
-    df_final.to_csv(f"{device}/Final_report_{Id_sim}.csv", index=False)
-    out_report = pd.read_csv(f"{device}/Final_report_{Id_sim}.csv")
-    print("\n", f"Max. mean error = {out_report['Mean error%'].max()}%")
-    print(
-        "====================================================================================================================================================="
-    )
+    df = pd.concat(all_dfs)
+    df.dropna(axis=0, inplace=True)
+    df["corner"] = "typical"
+    df = df[["device", "temp", "corner", "ic_measured", "ib_measured"]]
+
+    return df
+
+
+def ext_pnp_measured(dev_data_path, device, devices, dev_path):
+    # Read Data
+    df = pd.read_excel(dev_data_path)
+
+    all_dfs = []
+
+    loops = df["corners"].count()
+
+    for i in range(loops):
+
+        temp_range = int(loops / 4)
+        if i in range(0, temp_range):
+            temp = 25
+        elif i in range(temp_range, 2 * temp_range):
+            temp = -40
+        elif i in range(2 * temp_range, 3 * temp_range):
+            temp = 125
+        else:
+            temp = 175
+
+        tempr = []
+        dev = []
+        ic_meas = []
+        ib_meas = []
+
+        k = i
+        if i >= len(devices):
+            while k >= len(devices):
+                k = k - len(devices)
+
+        # Special case for 1st measured values
+
+        if i == 0:
+
+            idf_ic = df[["-vb ", "vc =-1", "vc =-2", "vc =-3"]].copy()
+            idf_ic.rename(
+                columns={
+                    "-vb ": "measured_base_volt",
+                    "vc =-1": "measured_ic_vcp_step1",
+                    "vc =-2": "measured_ic_vcp_step2",
+                    "vc =-3": "measured_ic_vcp_step3",
+                },
+                inplace=True,
+            )
+
+        else:
+
+            idf_ic = df[
+                ["-vb ", f"vc =-1.{2*i}", f"vc =-2.{2*i}", f"vc =-3.{2*i}"]
+            ].copy()
+            idf_ic.rename(
+                columns={
+                    "-vb ": "measured_base_volt",
+                    f"vc =-1.{2*i}": "measured_ic_vcp_step1",
+                    f"vc =-2.{2*i}": "measured_ic_vcp_step2",
+                    f"vc =-3.{2*i}": "measured_ic_vcp_step3",
+                },
+                inplace=True,
+            )
+
+        os.makedirs(f"{dev_path}/ic_measured", exist_ok=True)
+        idf_ic.to_csv(
+            f"{dev_path}/ic_measured/measured_{devices[k]}_t{temp}.csv"
+            )
+
+        idf_ib = df[
+            ["-vb ", f"vc =-1.{2*i+1}", f"vc =-2.{2*i+1}", f"vc =-3.{2*i+1}"]
+        ].copy()
+        idf_ib.rename(
+            columns={
+                "-vb ": "measured_base_volt",
+                f"vc =-1.{2*i+1}": "measured_ib_vcp_step1",
+                f"vc =-2.{2*i+1}": "measured_ib_vcp_step2",
+                f"vc =-3.{2*i+1}": "measured_ib_vcp_step3",
+            },
+            inplace=True,
+        )
+
+        os.makedirs(f"{dev_path}/ib_measured", exist_ok=True)
+        idf_ib.to_csv(
+            f"{dev_path}/ib_measured/measured_{devices[k]}_t{temp}.csv"
+            )
+
+        dev.append(devices[k])
+        tempr.append(temp)
+        ic_meas.append(
+            f"{dev_path}/ic_measured/measured_{devices[k]}_t{temp}.csv"
+            )
+        ib_meas.append(
+            f"{dev_path}/ib_measured/measured_{devices[k]}_t{temp}.csv"
+            )
+
+        sdf = {
+            "device": dev,
+            "temp": tempr,
+            "ic_measured": ic_meas,
+            "ib_measured": ib_meas,
+        }
+        sdf = pd.DataFrame(sdf)
+        all_dfs.append(sdf)
+
+    df = pd.concat(all_dfs)
+    df.dropna(axis=0, inplace=True)
+    df["corner"] = "typical"
+    df = df[["device", "temp", "corner", "ic_measured", "ib_measured"]]
+
+    return df
+
+
+def run_sim(char, dirpath, device, temp):
+    """ Run simulation at specific information and corner """
+
+    info = {}
+    info["device"] = device
+    info["temp"] = temp
+    dev = device.split("_")[0]
+
+    netlist_tmp = f"./device_netlists/{dev}.spice"
+
+    temp_str = "{:.1f}".format(temp)
+
+    netlist_path = f"{dirpath}/{dev}_netlists" \
+        + f"/netlist_{device}_t{temp_str}.spice"
+
+    result_path_ib = f"{dirpath}/ib_simulated/" \
+        + f"simulated_{device}_t{temp_str}.csv"
+    result_path_ic = f"{dirpath}/ic_simulated/" \
+        + f"simulated_{device}_t{temp_str}.csv"
+
+    for c in char:
+        os.makedirs(f"{dirpath}/{c}_simulated", exist_ok=True)
+
+    with open(netlist_tmp) as f:
+        tmpl = Template(f.read())
+        os.makedirs(f"{dirpath}/{dev}_netlists", exist_ok=True)
+
+        with open(netlist_path, "w") as netlist:
+            netlist.write(tmpl.render(device=device, temp=temp_str,))
+
+    # Running ngspice for each netlist
+    try:
+        call_simulator(netlist_path)
+        # Find bjt in csv
+        if find_bjt(result_path_ib):
+            bjt_simu_ib = result_path_ib
+            bjt_simu_ic = result_path_ic
+        else:
+            bjt_simu_ib = "None"
+            bjt_simu_ic = "None"
+    except Exception:
+        bjt_simu_ib = "None"
+        bjt_simu_ic = "None"
+
+    info["beta_sim_ib_unscaled"] = bjt_simu_ib
+    info["beta_sim_ic_unscaled"] = bjt_simu_ic
+
+    return info
+
+
+def run_sims(char, df, dirpath, num_workers=mp.cpu_count()):
+
+    results = []
+    with concurrent.futures.ThreadPoolExecutor(
+            max_workers=num_workers) as executor:
+        futures_list = []
+        for j, row in df.iterrows():
+            futures_list.append(
+                executor.submit(
+                    run_sim, char, dirpath, row["device"], row["temp"])
+            )
+
+        for future in concurrent.futures.as_completed(futures_list):
+            try:
+                data = future.result()
+                results.append(data)
+            except Exception as exc:
+                print("Test case generated an exception: %s" % (exc))
+
+    for c in char:
+        sf = glob.glob(
+            f"{dirpath}/{c}_simulated/*.csv"
+            )  # stored simulated data files
+
+        for i in range(len(sf)):
+            sdf = pd.read_csv(sf[i], header=None, delimiter=r"\s+",)
+
+            sweep = len(
+                pd.read_csv(glob.glob(f"{dirpath}/{c}_measured/*.csv")[1])
+                )
+
+            new_array = np.empty((sweep, 1 + int(sdf.shape[0] / sweep)))
+            new_array[:, 0] = sdf.iloc[:sweep, 0]
+            times = int(sdf.shape[0] / sweep)
+
+            for j in range(times):
+                new_array[:, (j + 1)] = sdf.iloc[j * sweep: (j + 1) * sweep, 1]
+
+            # Writing final simulated data 1
+            sdf = pd.DataFrame(new_array)
+            sdf.to_csv(
+                sf[i], index=False,
+            )
+
+            sdf.rename(
+                columns={
+                    0: "simulated_base_volt",
+                    1: f"simulated_{c}_vcp_step1",
+                    2: f"simulated_{c}_vcp_step2",
+                    3: f"simulated_{c}_vcp_step3",
+                },
+                inplace=True,
+            )
+
+            sdf.to_csv(sf[i])
+
+    df = pd.DataFrame(results)
+
+    df = df[["device", "temp", "beta_sim_ib_unscaled", "beta_sim_ic_unscaled"]]
+    df["beta_ib_sim"] = df["beta_sim_ib_unscaled"]
+    df["beta_ic_sim"] = df["beta_sim_ic_unscaled"]
+
+    return df
 
 
 def main():
@@ -342,55 +396,227 @@ def main():
     pd.set_option("max_colwidth", None)
     pd.set_option("display.width", 1000)
 
+    main_regr_dir = "bjt_beta_regr"
+
+    # bjt var.
+
     devices = ["npn", "pnp"]
-    list_devices = [
-        [
-            "npn_10p00x10p00",
-            "npn_05p00x05p00",
-            "npn_00p54x16p00",
-            "npn_00p54x08p00",
-            "npn_00p54x04p00",
-            "npn_00p54x02p00",
-        ],
-        ["pnp_10p00x00p42", "pnp_05p00x00p42", "pnp_10p00x10p00", "pnp_05p00x05p00"],
+
+    npn_devices = [
+        "npn_10p00x10p00",
+        "npn_05p00x05p00",
+        "npn_00p54x16p00",
+        "npn_00p54x08p00",
+        "npn_00p54x04p00",
+        "npn_00p54x02p00",
     ]
-    vb = ["vbp ", "-vb (V)"]
-    vc = ["vcp =", "vc =-"]
-    Id_sim = ["Ic", "Ib"]
-    sweep = 101
-    step = [1, 2, 3]
 
-    for i, device in enumerate(devices):
-        # Folder structure of measured values
-        dirpath = f"{device}"
-        if os.path.exists(dirpath) and os.path.isdir(dirpath):
-            shutil.rmtree(dirpath)
-        os.makedirs(f"{device}/measured_{Id_sim[0]}", exist_ok=False)
-        os.makedirs(f"{device}/measured_{Id_sim[1]}", exist_ok=False)
+    pnp_devices = [
+        "pnp_10p00x00p42",
+        "pnp_05p00x00p42",
+        "pnp_10p00x10p00",
+        "pnp_05p00x05p00",
+    ]
 
-        # From xlsx to csv
-        read_file = pd.read_excel(
-            f"../../180MCU_SPICE_DATA/BJT/bjt_{device}_beta_f.nl_out.xlsx"
+    char = ["ib", "ic"]
+
+    for i, dev in enumerate(devices):
+        dev_path = f"{main_regr_dir}/{dev}"
+
+        if os.path.exists(dev_path) and os.path.isdir(dev_path):
+            shutil.rmtree(dev_path)
+
+        os.makedirs(f"{dev_path}", exist_ok=False)
+
+        print("######" * 10)
+        print(f"# Checking Device {dev}")
+
+        print("\n")
+
+        # for c in char :
+
+        beta_data_files = glob.glob(
+            f"../../180MCU_SPICE_DATA/BJT/bjt_{dev}_beta_f.nl_out.xlsx"
         )
-        read_file.to_csv(f"{device}/{device}.csv", index=False, header=True)
+        if len(beta_data_files) < 1:
+            print("# Can't find diode file for device: {}".format(dev))
+            beta_file = ""
+        else:
+            beta_file = beta_data_files[0]
+        print("# bjt_beta data points file : ", beta_file)
 
-        # Folder structure of simulated values
-        os.makedirs(f"{device}/simulated_{Id_sim[0]}", exist_ok=False)
-        os.makedirs(f"{device}/error_{Id_sim[0]}", exist_ok=False)
-        os.makedirs(f"{device}/simulated_{Id_sim[1]}", exist_ok=False)
-        os.makedirs(f"{device}/error_{Id_sim[1]}", exist_ok=False)
+        if beta_file == "":
+            print(f"# No datapoints available for validation for device {dev}")
+            continue
 
-        # =========== Simulate ==============
-        ext_measured(device, vb[i], step, Id_sim, list_devices[i], vc[i])
+        if dev == "npn":
+            f = ext_npn_measured
+            list_dev = npn_devices
+        elif dev == "pnp":
+            f = ext_pnp_measured
+            list_dev = pnp_devices
 
-        ext_simulated(device, vb[i], step, sweep, Id_sim, list_devices[i], vc[i])
+        if beta_file != "":
+            meas_df = f(beta_file, dev, list_dev, dev_path)
+        else:
+            meas_df = []
 
-        # ============ Results =============
-        error_cal(device, vb[i], step, Id_sim[0], list_devices[i], vc[i])
-        error_cal(device, vb[i], step, Id_sim[1], list_devices[i], vc[i])
+        meas_len = len(
+            pd.read_csv(glob.glob(f"{dev_path}/ic_measured/*.csv")[1])
+            )
+
+        print(
+            f"# Device {dev} number of measured_datapoints : ",
+            len(meas_df) * meas_len
+        )
+
+        sim_df = run_sims(char, meas_df, dev_path, 3)
+
+        sim_len = len(
+            pd.read_csv(glob.glob(f"{dev_path}/{char[1]}_simulated/*.csv")[1])
+        )
+
+        print(
+            f"# Device {dev} number of simulated datapoints : ",
+            len(sim_df) * sim_len
+        )
+
+        # compare section
+
+        merged_df = meas_df.merge(sim_df, on=["device", "temp"], how="left")
+
+        merged_all = []
+        for c in char:
+
+            merged_dfs = []
+
+            for i in range(len(merged_df)):
+
+                measured_data = pd.read_csv(merged_df[f"{c}_measured"][i])
+                simulated_data = pd.read_csv(merged_df[f"beta_{c}_sim"][i])
+
+                result_data = simulated_data.merge(measured_data, how="left")
+
+                result_data["corner"] = "typical"
+                result_data["device"] = (
+                    merged_df[f"{c}_measured"][i]
+                    .split("/")[-1]
+                    .split("d_")[1]
+                    .split("_t")[0]
+                )
+                result_data["temp"] = (
+                    merged_df[f"{c}_measured"][i]
+                    .split("/")[-1]
+                    .split("_")[3]
+                    .split("t")[1]
+                    .split(".")[0]
+                )
+
+                result_data["step1_error"] = (
+                    np.abs(
+                        result_data[f"simulated_{c}_vcp_step1"]
+                        - result_data[f"measured_{c}_vcp_step1"]
+                    )
+                    * 100.0
+                    / result_data[f"measured_{c}_vcp_step1"]
+                )
+
+                result_data["step2_error"] = (
+                    np.abs(
+                        result_data[f"simulated_{c}_vcp_step2"]
+                        - result_data[f"measured_{c}_vcp_step2"]
+                    )
+                    * 100.0
+                    / result_data[f"measured_{c}_vcp_step2"]
+                )
+
+                result_data["step3_error"] = (
+                    np.abs(
+                        result_data[f"simulated_{c}_vcp_step3"]
+                        - result_data[f"measured_{c}_vcp_step3"]
+                    )
+                    * 100.0
+                    / result_data[f"measured_{c}_vcp_step3"]
+                )
+
+                result_data["error"] = (
+                    np.abs(
+                        result_data["step1_error"]
+                        + result_data["step2_error"]
+                        + result_data["step3_error"]
+                    )
+                    / 3
+                )
+
+                result_data = result_data[
+                    [
+                        "device",
+                        "temp",
+                        "corner",
+                        "measured_base_volt",
+                        f"measured_{c}_vcp_step1",
+                        f"measured_{c}_vcp_step2",
+                        f"measured_{c}_vcp_step3",
+                        f"simulated_{c}_vcp_step1",
+                        f"simulated_{c}_vcp_step2",
+                        f"simulated_{c}_vcp_step3",
+                        "step1_error",
+                        "step2_error",
+                        "step3_error",
+                        "error",
+                    ]
+                ]
+
+                merged_dfs.append(result_data)
+
+            merged_out = pd.concat(merged_dfs)
+
+            merged_out.to_csv(
+                f"{dev_path}/error_analysis_{c}.csv", index=False
+                )
+
+            merged_all.append(merged_out)
+
+        merged_all = pd.concat(merged_all)
+
+        if merged_all["error"].min() > 100:
+            min_error = 100
+        else:
+            min_error = merged_all["error"].min()
+
+        if merged_all["error"].max() > 100:
+            max_error = 100
+        else:
+            max_error = merged_all["error"].max()
+
+        if merged_all["error"].mean() > 100:
+            mean_error = 100
+        else:
+            mean_error = merged_all["error"].mean()
+
+        print(
+            "# Device {} min error: {:.2f}"
+            .format(
+                dev, min_error
+            ),
+            ", max error: {:.2f}, mean error {:.2f}"
+            .format(
+                max_error, mean_error
+                )
+        )
+
+        if merged_out["error"].max() < PASS_THRESH:
+            print("# Device {} has passed regression.".format(dev))
+        else:
+            print(
+                "# Device {} has failed regression. Needs more analysis."
+                .format(dev)
+                )
+
+        print("\n\n")
 
 
-# ================================================================
+# # ================================================================
 # -------------------------- MAIN --------------------------------
 # ================================================================
 
@@ -400,7 +626,7 @@ if __name__ == "__main__":
     arguments = docopt(__doc__, version="comparator: 0.1")
     workers_count = (
         os.cpu_count() * 2
-        if arguments["--num_cores"] == None
+        if arguments["--num_cores"] is None
         else int(arguments["--num_cores"])
     )
