+++ models/ngspice/testing/regression/mimcap_c/models_regression.py
@@ -1,4 +1,4 @@
-# Copyright 2022 Efabless Corporation
+# Copyright 2022 GlobalFoundries PDK Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,14 +13,14 @@
 # limitations under the License.
 """
 Usage:
-  models_regression.py [--num_cores=<num>]
+  model_reg.py [--num_cores=<num>]
 
   -h, --help             Show help text.
   -v, --version          Show version.
   --num_cores=<num>      Number of cores to be used by simulator
 """
 
-from re import T
+from unittest.mock import DEFAULT
 from docopt import docopt
 import pandas as pd
 import numpy as np
@@ -28,170 +28,155 @@ import os
 from jinja2 import Template
 import concurrent.futures
 import shutil
+import multiprocessing as mp
+
+import subprocess
+import glob
+
 import warnings
 
 warnings.simplefilter(action="ignore", category=FutureWarning)
 
+DEFAULT_TEMP = 25.0
+PASS_THRESH = 2.0
+
+def find_mimcap(filename):
+    """
+    Find mimcap in log
+    """
+    cmd = 'grep "cv" {} | head -n 1'.format(filename)
+    process = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE)
+    return float(process.communicate()[0][:-1].decode("utf-8").split("=")[1])
 
 def call_simulator(file_name):
     """Call simulation commands to perform simulation.
     Args:
         file_name (str): Netlist file name.
     """
-    os.system(f"ngspice -b -a {file_name} -o {file_name}.log > {file_name}.log")
+    return os.system(f"ngspice -b -a {file_name} -o {file_name}.log > {file_name}.log")
 
+def ext_measured(dev_data_path, device, corners):
+    # Read Data
+    df = pd.read_excel(dev_data_path)
+    
+    length = []
+    width = []
 
-def ext_measured(device, vn, d_in, cv_sim, corner, start):
+    for i in range(len(df)):
 
-    # Get dimensions used for each device
-    dirpath = f"{device}_{cv_sim}_{corner}"
+        a_str = df["Unnamed: 2"][i]  # area_string parameter in the format "mim_1p5fF(100u x100u )"
+        if type(a_str) == str :
 
-    # Extracting measured values for each W & L
-    for i in range(start, 4 + start):
-        if i == 0 + start:
-            width = 100
-            length = 100
-        if i == 1 + start:
-            width = 5
-            length = 5
-        if i == 2 + start:
-            width = 100
-            length = 5
-        if i == 3 + start:
-            width = 5
-            length = 100
-
-        # measured cv
-        col_list = [f"{vn}", f"{d_in}"]
-        df_measured = pd.read_csv(f"{dirpath}/{device}.csv", usecols=col_list)
-        df_measured.loc[i:i].to_csv(
-            f"{dirpath}/measured_{cv_sim}/{i-start}_measured_w{width}_l{length}.csv",
-            index=False,
+            length.append(float(a_str.split("(")[1].split("x")[0].split("u")[0]))
+            width.append(float(a_str.split("(")[1].split("x")[1].split("u")[0]))
+        
+        else :
+            length.append(a_str)
+            width.append(a_str)
+    
+    all_dfs = []
+    for corner in corners:
+        idf = df[[f"mimcap_{corner}"]].copy()
+        idf.rename(
+            columns={
+                f"mimcap_{corner}": "mimcap_measured",
+            },
+            inplace=True,
         )
+        idf["corner"] = corner
+        idf["length"] = length
+        idf["width"] = width
+        all_dfs.append(idf)
+
+    df = pd.concat(all_dfs)
+    df["temp"] = DEFAULT_TEMP
+    df["device"] = device
+    df.dropna(axis=0, inplace=True)
+    df = df[["device", "corner", "length", "width", "temp", "mimcap_measured"]]
+    return df
+
+def run_sim(dirpath, device, length, width, corner, temp=25):
+    """ Run simulation at specific information and corner """
+    netlist_tmp = "./device_netlists/mimcap.spice"
 
 
-def ext_simulated(device, vn, d_in, cv_sim, corner, start):
+    info = {}
+    info["device"] = device
+    info["corner"] = corner
+    info["temp"] = temp
+    info["width"] = width
+    info["length"] = length
 
-    # Get dimensions used for each device
-    dirpath = f"{device}_{cv_sim}_{corner}"
-    netlist_tmp = f"./device_netlists/mimcap.spice"
+    width_str = "{:.0f}".format(width)
+    length_str = "{:.0f}".format(length)
+    temp_str = "{:.1f}".format(temp)
 
-    # Extracting measured values for each W & L
-    for i in range(start, 4 + start):
-        if i == 0 + start:
-            width = 100
-            length = 100
-        if i == 1 + start:
-            width = 5
-            length = 5
-        if i == 2 + start:
-            width = 100
-            length = 5
-        if i == 3 + start:
-            width = 5
-            length = 100
+    netlist_path = f"{dirpath}/{device}_netlists/netlist_w{width_str}_l{length_str}_t{temp_str}_{corner}.spice"
 
-        with open(netlist_tmp) as f:
-            tmpl = Template(f.read())
-            os.makedirs(f"{dirpath}/{device}_netlists_{cv_sim}", exist_ok=True)
-            with open(
-                f"{dirpath}/{device}_netlists_{cv_sim}/{i-start}_{device}_netlist_w{width}_l{length}.spice",
-                "w",
-            ) as netlist:
-                netlist.write(
-                    tmpl.render(
-                        device=device, width=width, length=length, corner=corner
-                    )
+    with open(netlist_tmp) as f:
+        tmpl = Template(f.read())
+        os.makedirs(f"{dirpath}/{device}_netlists", exist_ok=True)
+
+        with open(netlist_path, "w") as netlist:
+            netlist.write(
+                tmpl.render(
+                    device=device,
+                    width=width_str,
+                    length=length_str,
+                    corner=corner,
+                    temp=temp_str,
                 )
-            netlist_path = f"{dirpath}/{device}_netlists_{cv_sim}/{i-start}_{device}_netlist_w{width}_l{length}.spice"
-            # Running ngspice for each netlist
-            with concurrent.futures.ProcessPoolExecutor(
-                max_workers=workers_count
-            ) as executor:
-                executor.submit(call_simulator, netlist_path)
-
-            # Writing simulated data
-            df_simulated = pd.read_csv(
-                f"{dirpath}/{device}_netlists_{cv_sim}/{i-start}_{device}_netlist_w{width}_l{length}.spice.log"
-            )
-            clean_data = (
-                str(df_simulated.loc[4])
-                .replace("Compatibility modes selected: hs", "")
-                .replace("\nName: 4, dtype: object", "")
-                .split("=")
-            )
-            df_clean_ = {vn: [f"moscap_{corner}"], d_in: [clean_data[1]]}
-            df_clean = pd.DataFrame(df_clean_)
-            df_clean.to_csv(
-                f"{dirpath}/simulated_{cv_sim}/{i-start}_simulated_w{width}_l{length}.csv",
-                index=False,
             )
 
+    # Running ngspice for each netlist
+    try:
+        call_simulator(netlist_path)
+        # Find mimcap in log
+        try:
+            mim = find_mimcap(f"{netlist_path}.log")
+            
+        except Exception as e:
+            mim = 0.0
+        
+    except Exception as e:
+        mim = 0.0
+    
+    info["mim_sim_unscaled"] = mim
 
-def error_cal(device, vn, d_in, Id_sim, corner, start):
+    return info
 
-    # Get dimensions used for each device
-    dirpath = f"{device}_{Id_sim}_{corner}"
-    df_final = pd.DataFrame()
-    for i in range(start, 4 + start):
-        if i == 0 + start:
-            width = 100
-            length = 100
-        if i == 1 + start:
-            width = 5
-            length = 5
-        if i == 2 + start:
-            width = 100
-            length = 5
-        if i == 3 + start:
-            width = 5
-            length = 100
+def run_sims(df, dirpath, num_workers=mp.cpu_count()):
 
-        measured = pd.read_csv(
-            f"{dirpath}/measured_{Id_sim}/{i-start}_measured_w{width}_l{length}.csv"
-        )
-        simulated = pd.read_csv(
-            f"{dirpath}/simulated_{Id_sim}/{i-start}_simulated_w{width}_l{length}.csv"
-        )
+    results = []
+    with concurrent.futures.ThreadPoolExecutor(max_workers=num_workers) as executor:
+        futures_list = []
+        for j, row in df.iterrows():
+            futures_list.append(
+                executor.submit(
+                    run_sim,
+                    dirpath,
+                    row["device"],
+                    row["length"],
+                    row["width"],
+                    row["corner"],
+                    row["temp"],
+                )
+            )
 
-        error_1 = round(
-            100
-            * abs(
-                (abs(measured.iloc[:, 1]) - abs(simulated.iloc[:, 1]))
-                / abs(measured.iloc[:, 1])
-            ),
-            8,
-        )
+        for future in concurrent.futures.as_completed(futures_list):
+            try:
+                data = future.result()
+                results.append(data)
+            except Exception as exc:
+                print("Test case generated an exception: %s" % (exc))
 
-        df_error = pd.DataFrame(data=[measured.iloc[:, 0], error_1]).transpose()
-        df_error.to_csv(
-            f"{dirpath}/error_{Id_sim}/{i-start}_{device}_error_w{width}_l{length}.csv",
-            index=False,
-        )
+    df = pd.DataFrame(results)
+    df = df[
+        ["device", "corner", "length", "width", "temp","mim_sim_unscaled"]
+    ]
 
-        # Mean error
-        mean_error = df_error[f"{d_in}"].mean()
-        # Max error
-        max_error = df_error[f"{d_in}"].max()
-
-        df_final_ = {
-            "Run no.": f"{i-start}",
-            "Device name": f"{dirpath}",
-            "Width": f"{width}",
-            "Length": f"{length}",
-            "Simulated_Val": f"{Id_sim}",
-            "Mean error%": f'{"{:.2f}".format(mean_error)}',
-            "Max error%": f'{"{:.2f}".format(max_error)} ',
-        }
-        df_final = df_final.append(df_final_, ignore_index=True)
-    # Max mean error
-    print(df_final)
-    df_final.to_csv(f"{dirpath}/Final_report_{Id_sim}.csv", index=False)
-    out_report = pd.read_csv(f"{dirpath}/Final_report_{Id_sim}.csv")
-    print("\n", f"Max. mean error = {out_report['Mean error%'].max()}%")
-    print(
-        "====================================================================================================================================================="
-    )
+    df["mim_sim"] = df["mim_sim_unscaled"]
+    return df
 
 
 def main():
@@ -202,40 +187,80 @@ def main():
     pd.set_option("max_colwidth", None)
     pd.set_option("display.width", 1000)
 
-    # mim
-    corners = ["ss", "typical", "ff"]
-    devices = [
+    main_regr_dir = "mimcap_regr"
+
+    # mimcap var.
+    corners = ["typical", "ff", "ss"]
+
+    devices = [ 
         "cap_mim_1f5_m2m3_noshield",
         "cap_mim_1f0_m3m4_noshield",
         "cap_mim_2f0_m4m5_noshield",
     ]
 
-    measure = ["cv", "corners", "CV (fF)"]
-    start = 0
-    for corner in corners:
-        for device in devices:
-            # Folder structure of measured values
-            cv_sim, cap_vn, cap_in = measure[0], measure[1], measure[2]
-            dirpath = f"{device}_{cv_sim}_{corner}"
-            if os.path.exists(dirpath) and os.path.isdir(dirpath):
-                shutil.rmtree(dirpath)
-            os.makedirs(f"{dirpath}/measured_{cv_sim}", exist_ok=False)
+    for i, dev in enumerate(devices):
+        dev_path = f"{main_regr_dir}/{dev}"
 
-            # From xlsx to csv
-            read_file = pd.read_excel(
-                f"../../180MCU_SPICE_DATA/Cap/mimcap_fc.nl_out.xlsx"
+        if os.path.exists(dev_path) and os.path.isdir(dev_path):
+            shutil.rmtree(dev_path)
+
+        os.makedirs(f"{dev_path}", exist_ok=False)
+
+        print("######" * 10)
+        print(f"# Checking Device {dev}")
+
+        mim_data_files = glob.glob(
+            f"../../180MCU_SPICE_DATA/Cap/mimcap_fc.nl_out.xlsx"
+        )
+        if len(mim_data_files) < 1:
+            print("# Can't find mimcap file for device: {}".format(dev))
+            mim_file = ""
+        else:
+            mim_file = mim_data_files[0]
+        print("# mimcap data points file : ", mim_file)
+
+        if mim_file == "" :
+            print(f"# No datapoints available for validation for device {dev}")
+            continue
+            
+        if mim_file != "":
+            meas_df = ext_measured(mim_file, dev, corners)
+        else:
+            meas_df = []
+
+        print("# Device {} number of measured_datapoints : ".format(dev), len(meas_df))
+
+        sim_df = run_sims(meas_df, dev_path, 3)
+        print("# Device {} number of simulated datapoints : ".format(dev), len(sim_df))
+
+        merged_df = meas_df.merge(
+            sim_df, on=["device", "corner", "length", "width", "temp"], how="left"
+        )
+
+        merged_df["error"] = (
+            np.abs(merged_df["mim_sim"] - merged_df["mimcap_measured"])
+            * 100.0
+            / merged_df["mimcap_measured"]
+        )
+
+        merged_df.to_csv(f"{dev_path}/error_analysis.csv", index=False)
+
+        print(
+            "# Device {} min error: {:.2f} , max error: {:.2f}, mean error {:.2f}".format(
+                dev,
+                merged_df["error"].min(),
+                merged_df["error"].max(),
+                merged_df["error"].mean(),
             )
-            read_file.to_csv(f"{dirpath}/{device}.csv", index=False, header=True)
+        )
 
-            # Folder structure of simulated values
-            os.makedirs(f"{dirpath}/simulated_{cv_sim}", exist_ok=False)
-            os.makedirs(f"{dirpath}/error_{cv_sim}", exist_ok=False)
+        if merged_df["error"].max() < PASS_THRESH:
+            print("# Device {} has passed regression.".format(dev))
+        else:
+            print("# Device {} has failed regression. Needs more analysis.".format(dev))
 
-            ext_measured(device, cap_vn, cap_in, cv_sim, corner, start)
-            ext_simulated(device, cap_vn, cap_in, cv_sim, corner, start)
-            error_cal(device, cap_vn, cap_in, cv_sim, corner, start)
-            start = start + 4
-        start = start + 24
+        print("\n\n")
+        
 
 
 # # ================================================================
@@ -254,3 +279,5 @@ if __name__ == "__main__":
 
     # Calling main function
     main()
+
+
